+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the "Project Timer" extension will be documented in this file.
 
+## [Unreleased]
+### Fixed
+- Fixed cannot get git remote url correctly.
+
 ## [0.2.1] - 2026-2-21
 ### Added
 - Implemented MatchInfo, Calculator, and Config caching to reduce CPU overhead.